change data type for provinces

+++ db/migrate/20180402023111_create_provinces.rb
@@ -2,9 +2,9 @@ class CreateProvinces < ActiveRecord::Migration[5.1]
   def change
     create_table :provinces do |t|
       t.string :name
-      t.decimal :pst_rate
-      t.decimal :gst_rate
-      t.decimal :hst_rate
+      t.integer :pst_rate
+      t.integer :gst_rate
+      t.integer :hst_rate
 
       t.timestamps
     end
